Add Poppler::Version in its own file
